+++ CHANGELOG.md
@@ -6,12 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.9.0] - 2025-07-20
+
 ### Added
 
 - Ability to set custom project icons (and an alternative if the project icon is not available)
 
   - The icon for each project can be configured inside the "Discord Rich Presence" -> "Project" settings menu 
-
 - `{project_repo_branch}` variable that shows the name of the project's branch
 - "Reset settings to defaults" button
 
@@ -267,7 +268,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release
 
-[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.8.0...HEAD
+[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.9.0...HEAD
+[1.9.0]: https://github.com/pandier/intellij-discord-rp/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/pandier/intellij-discord-rp/compare/v1.7.4...v1.8.0
 [1.7.4]: https://github.com/pandier/intellij-discord-rp/compare/v1.7.3...v1.7.4
 [1.7.3]: https://github.com/pandier/intellij-discord-rp/compare/v1.7.2...v1.7.3